// lib/generated/prisma/models/GameList.ts
export type GameListModel = runtime.Types.Result.DefaultSelection<Prisma.$GameListPayload>
// ...
export type $GameListPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "GameList"
  objects: {
    entries: Prisma.$GameListEntryPayload<ExtArgs>[]
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    type: $Enums.GameListType
    name: string
    slug: string | null
    description: string | null
    image: string | null
    background: string | null
    color: string | null
    accentColor: string | null
    privacy: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["gameList"]>
  composites: {}
}